+++ beekeeper-core/src/test/java/com/expediagroup/beekeeper/core/repository/HousekeepingMetadataRepositoryTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2019-2020 Expedia, Inc.
+ * Copyright (C) 2019-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -44,6 +44,8 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.support.AnnotationConfigContextLoader;
 
+import com.google.common.collect.Lists;
+
 import com.expediagroup.beekeeper.core.TestApplication;
 import com.expediagroup.beekeeper.core.model.HousekeepingMetadata;
 
@@ -83,7 +85,7 @@ public class HousekeepingMetadataRepositoryTest {
 
     housekeepingMetadataRepository.save(table);
 
-    List<HousekeepingMetadata> tables = housekeepingMetadataRepository.findAll();
+    List<HousekeepingMetadata> tables = Lists.newArrayList(housekeepingMetadataRepository.findAll());
     assertThat(tables.size()).isEqualTo(1);
     HousekeepingMetadata savedTable = tables.get(0);
     assertThat(savedTable.getPath()).isEqualTo(PATH);
@@ -108,7 +110,7 @@ public class HousekeepingMetadataRepositoryTest {
     savedTable.setCleanupAttempts(savedTable.getCleanupAttempts() + 1);
     housekeepingMetadataRepository.save(savedTable);
 
-    List<HousekeepingMetadata> tables = housekeepingMetadataRepository.findAll();
+    List<HousekeepingMetadata> tables = Lists.newArrayList(housekeepingMetadataRepository.findAll());
     assertThat(tables.size()).isEqualTo(1);
     HousekeepingMetadata updatedTable = tables.get(0);
     assertThat(updatedTable.getHousekeepingStatus()).isEqualTo(DELETED);
@@ -141,7 +143,7 @@ public class HousekeepingMetadataRepositoryTest {
   public void timezone() {
     HousekeepingMetadata path = createPartitionedEntityHousekeepingTable();
     housekeepingMetadataRepository.save(path);
-    List<HousekeepingMetadata> tables = housekeepingMetadataRepository.findAll();
+    List<HousekeepingMetadata> tables = Lists.newArrayList(housekeepingMetadataRepository.findAll());
     assertThat(tables.size()).isEqualTo(1);
 
     HousekeepingMetadata savedTable = tables.get(0);
@@ -370,7 +372,7 @@ public class HousekeepingMetadataRepositoryTest {
       String databaseName,
       String tableName,
       String partitionName) {
-    return new HousekeepingMetadata.Builder()
+    return HousekeepingMetadata.builder()
         .path(PATH)
         .databaseName(databaseName)
         .tableName(tableName)