+++ CHANGELOG.md
@@ -3,7 +3,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-## [3.2.0] - 2021-07-14
+
+## [3.4.0] - 2021-08-18
+### Added
+- Added a `GET /unreferenced-paths` endpoint to the `beekeeper-api`.
+
+## [3.3.0] - 2021-08-12
+### Added
+- Added a `GET /metadata` endpoint to the `beekeeper-api`.
+### Changed
+- Updated `eg-oss-parent` to version `2.4.0` (was `2.3.2`).
+- Updated `snakeyaml` to version `1.27` (was `1.24`).
+- Updated `mockito.version` to version `3.11.2` (was `3.9.0`).
+
+## [3.2.0] - 2021-07-21
 ### Added
 - Added `beekeeper-api` module.
 ### Changed